[BUGFIX] Fix rendering of marketing data
This patch fixes the rendering of the marketing data charts that are
broken after the resturcturing.

The getter for the marketingPageFunnel will now always return an array
to prevent warnings in other places. Additinally a new getter was added
that returns the UID of the last visited page.

The ReportingMarketingBe template was restructured and is now displaying
charts of the new marketing properties.

Additinally a condition in the ListBe template is fixed and debug output
is removed.

WIP: Since it would be good to display the page names in the charts too
the page title needs to be retrieved already in the Div class.

Releases: 2.1
Change-Id: I23c4f2a8bb56f05900fbb10411e0e8ce6a1bedb4
Reviewed-on: https://review.typo3.org/27659
Reviewed-by: Alex Kellner
Tested-by: Alex Kellner