[BUGFIX] Prevent NULL value for marker column
Since TYPO3 6.2 allows inserting NULL values in the database you will
get an error when you switch the field type before entering a name
because the marker value is NULL instead of an empty string.

This patch adds a typecast to the initialization of the marker column
to prevent this database error.

Releases: 2.1
Change-Id: Id79fbfe6ca43125cd5f901f2b0ef6573ce7a5f0e
Reviewed-on: https://review.typo3.org/27667
Reviewed-by: Alex Kellner
Tested-by: Alex Kellner