[BUGFIX] Property mapping config for TYPO3 6.1
The property mapper was updated in TYPO3 / Extbase 6.1 (see #47182)
it is required to explicitly specify the properties that are allowed
to be mapped.

To prevent mapping errors fields for which the UID can not be
determined (e.g. the secondary password field) are skipped during the
creation of the fields array.

Additionally the check for upload data was improved. If no file is
uploaded the 'tmp_name' array key exists but is empty. To prevent a
mapping error by using the file data array as value the check was
changed from empty() to isset().

If the value is still an array (e.g. for checkboxes) the data will
be serialized. Desearialization is already handled by the answer
domain model.

Resolves: #55993
Change-Id: Ia44c2ed5d735e9f20c7a270275f23313411fc501