[FEATURE] Make forms localizable using IRRE
In order to localize form labels and prefill values while preserving
the ease of use of IRRE, the localization configuration must be
changed.

The l10n_mode of the relations forms->pages and pages->fields must
be changed to mergeIfNotBlank. Additionally, all fields that should
be configured in the original language version (such as Mandatory,
Validation etc.) should be readonly fields in the localized version.

The feature "Enable translating of whole forms using IRRE" must be
enabled in the Extension Manager to activate the configuration.

The feature only works if the conflicting feature "Enable Merge
for l10n_mode" is disabled

Releases: 3.0.0-F